Track sort direction in SortGroupClause
Functions make_pathkey_from_sortop() and
transformWindowDefinitions() which receive a SortGroupClause were
determining the sort order (ascending vs. descending) by comparing
that structure's operator strategy to BTLessStrategyNumber,
but could just as easily have gotten it from the SortGroupClause
object if it had such a field, so add one.  This reduces the number
of places that hardcode the assumption that the strategy refers
specifically to a btree strategy, rather than some other index AM's
operators.