Allow non-btree indexes to participate in mergejoin
Rather than hardcoding a requirement that the index be of type
btree, just require that it provide ordering support.

The regression test expected output for xtree changes, due to a
pre-existing test that creates a broken opfamily now causing
trouble.  It is good and proper that the code should complain,
though the Oid of the operator being present in the complaint
would cause the regression test containing the complaint to be
unstable, so change the complaint language to name the operator
rather than giving its Oid.