Replace various references to btree strategies
A number of places in the code still had references to btree
strategies where not appropriate, so fix those.  Leave untouched
such strategy numbers in nbtree, spgist, and brin indexes, and also
in table partitioning.  Hardcoded hash and btree references in table
partitioning are part of the design of table partitioning; their
replacement, if even a good idea, is beyond the scope of this patch
series.