Make index insert compatible with outside callers
We split aminsert method to aminsert and aminsertextended.

aminsert is a method for indexes implemented in other extensions, it
accepts ItemPointer tupleid.

aminsertextended is for internal Postgres indexes and Orioledb, it
accepts Datum tupleid. They are not supposed to call aminsert method, so
that it is set NULL for them. We can not rely that extensions are aware
of aminsertextended, so index_insert() calls aminsert if it's not NULL
preferentially.

Signature of index_insert() is reverted so that it could be called by other
extensions. Datum tupleid is confined inside index_insert method.